予算戦略の追加とそれに合わせてリファクタリング

+++ internal/infra/kabu/types.go
@@ -68,7 +68,7 @@ type Position struct {
 	ExecutionID    string  `json:"ExecutionID"`    // 約定番号（決済指定時に使う）
 	Symbol         string  `json:"Symbol"`         // 銘柄コード (例: "7012")
 	SymbolName     string  `json:"SymbolName"`     // 銘柄名
-	LeavesQty      float64 `json:"LeavesQty"`      // 残数量（いま決済できる株数）
+	LeavesQty      uint32  `json:"LeavesQty"`      // 残数量（いま決済できる株数）
 	HoldQty        float64 `json:"HoldQty"`        // 拘束数量（すでに売り注文を出して待機中の株数）
 	Price          float64 `json:"Price"`          // 建値（平均取得単価） ★0.2%計算の基準！
 	CurrentPrice   float64 `json:"CurrentPrice"`   // 現在値
